@@ -12,9 +12,10 @@
         }
     ],
     "require": {
-        "php": ">=8.1",
-        "symfony/http-client": "^5.4|^6.0",
-        "symfony/notifier": "^6.2.7"
+        "php": ">=8.2",
+        "symfony/http-client": "^6.4|^7.0",
+        "symfony/mime": "^6.4|^7.0",
+        "symfony/notifier": "^6.4|^7.0"
     },
     "autoload": {
         "psr-4": { "Proxykon\\Component\\Notifier\\Bridge\\Telegram\\": "" },
