@@ -11,6 +11,7 @@
 
 namespace Proxykon\Component\Notifier\Bridge\Telegram;
 
+use Symfony\Component\Notifier\Exception\MultipleExclusiveOptionsUsedException;
 use Symfony\Component\Notifier\Exception\TransportException;
 use Symfony\Component\Notifier\Exception\UnsupportedMessageTypeException;
 use Symfony\Component\Notifier\Message\ChatMessage;
@@ -34,6 +35,20 @@ final class TelegramTransport extends AbstractTransport
 
     private string $token;
     private ?string $chatChannel;
+
+    private const EXCLUSIVE_OPTIONS = [
+        'message_id',
+        'callback_query_id',
+        'photo',
+        'location',
+        'audio',
+        'document',
+        'video',
+        'animation',
+        'venue',
+        'contact',
+        'sticker',
+    ];
 
     public function __construct(#[\SensitiveParameter] string $token, ?string $channel = null, ?HttpClientInterface $client = null, ?EventDispatcherInterface $dispatcher = null)
     {
@@ -68,23 +83,38 @@ final class TelegramTransport extends AbstractTransport
         }
 
         $options = $message->getOptions()?->toArray() ?? [];
+        $optionsContainer = 'json';
         $options['chat_id'] ??= $message->getRecipientId() ?: $this->chatChannel;
-        $options['text'] = $message->getSubject();
+        $text = $message->getSubject();
 
         if (!isset($options['parse_mode']) || TelegramOptions::PARSE_MODE_MARKDOWN_V2 === $options['parse_mode']) {
             $options['parse_mode'] = TelegramOptions::PARSE_MODE_MARKDOWN_V2;
-            $options['text'] = preg_replace('/([_*\[\]()~`>#+\-=|{}.!\\\\])/', '\\\\$1', $message->getSubject());
+            $text = preg_replace('/([_*\[\]()~`>#+\-=|{}.!\\\\])/', '\\\\$1', $text);
         }
 
-        if (isset($options['photo'])) {
-            $options['caption'] = $options['text'];
-            unset($options['text']);
+        if (isset($options['upload'])) {
+            foreach ($options['upload'] as $option => $document) {
+                $options[$option] = fopen($document['path'], 'r');
+                $fNamePart = explode('.', $document['fileName']);
+                stream_context_set_option($options[$option], 'http', 'filename', $document['fileName']);
+                stream_context_set_option($options[$option], 'http', 'content_type', $document['mimeType']);
+            }
+            $optionsContainer = 'body';
+            unset($options['upload']);
         }
 
-        $endpoint = sprintf('https://%s/bot%s/%s', $this->getEndpoint(), $this->token, $this->getPath($options));
+        $messageOption = $this->getTextOption($options);
+        if (null !== $messageOption) {
+            $options[$messageOption] = $text;
+        }
+        $method = $this->getPath($options);
+        $this->ensureExclusiveOptionsNotDuplicated($options);
+        $options = $this->expandOptions($options, 'contact', 'location', 'venue');
+
+        $endpoint = sprintf('https://%s/bot%s/%s', $this->getEndpoint(), $this->token, $method);
 
         $response = $this->client->request('POST', $endpoint, [
-            'json' => array_filter($options),
+            $optionsContainer => array_filter($options),
         ]);
 
         try {
@@ -115,6 +145,14 @@ final class TelegramTransport extends AbstractTransport
             isset($options['message_id']) => 'editMessageText',
             isset($options['callback_query_id']) => 'answerCallbackQuery',
             isset($options['photo']) => 'sendPhoto',
+            isset($options['location']) => 'sendLocation',
+            isset($options['audio']) => 'sendAudio',
+            isset($options['document']) => 'sendDocument',
+            isset($options['video']) => 'sendVideo',
+            isset($options['animation']) => 'sendAnimation',
+            isset($options['venue']) => 'sendVenue',
+            isset($options['contact']) => 'sendContact',
+            isset($options['sticker']) => 'sendSticker',
             default => 'sendMessage',
         };
     }
@@ -126,5 +164,44 @@ final class TelegramTransport extends AbstractTransport
             isset($options['callback_query_id']) => 'answer callback query',
             default => 'post',
         };
+    }
+
+    private function getTextOption(array $options): ?string
+    {
+        return match (true) {
+            isset($options['photo']) => 'caption',
+            isset($options['audio']) => 'caption',
+            isset($options['document']) => 'caption',
+            isset($options['video']) => 'caption',
+            isset($options['animation']) => 'caption',
+            isset($options['sticker']) => null,
+            isset($options['location']) => null,
+            isset($options['venue']) => null,
+            isset($options['contact']) => null,
+            default => 'text',
+        };
+    }
+
+    private function expandOptions(array $options, string ...$optionsForExpand): array
+    {
+        foreach ($optionsForExpand as $optionForExpand) {
+            if (isset($options[$optionForExpand])) {
+                if (\is_array($options[$optionForExpand])) {
+                    $options = array_merge($options, $options[$optionForExpand]);
+                }
+                unset($options[$optionForExpand]);
+            }
+        }
+
+        return $options;
+    }
+
+    private function ensureExclusiveOptionsNotDuplicated(array $options): void
+    {
+        $usedOptions = array_keys($options);
+        $usedExclusiveOptions = array_intersect($usedOptions, self::EXCLUSIVE_OPTIONS);
+        if (\count($usedExclusiveOptions) > 1) {
+            throw new MultipleExclusiveOptionsUsedException($usedExclusiveOptions, self::EXCLUSIVE_OPTIONS);
+        }
     }
 }
